+++ man-roxygen/seealso_task_generator.R
@@ -1,0 +1,7 @@
+#' @seealso
+#'
+#' * [Dictionary][mlr3misc::Dictionary] of [TaskGenerators][mlr3::TaskGenerator]: [mlr3::mlr_task_generators]
+#' * `as.data.table(mlr_task_generators)` for a table of available [TaskGenerators][mlr3::TaskGenerator] in the
+#'   running session (depending on the loaded packages).
+#'
+#' @family TaskGenerator